What was done: Extracted the duplicated validateAndResolveParams method into SharedTrainLogic (where the other shared training helpers already live). Both IsolationForest.scala   and ExtendedIsolationForest.scala now call the single shared implementation, passing $(maxFeatures) and $(maxSamples) as arguments.
  Files changed:
  - core/SharedTrainLogic.scala — added validateAndResolveParams(dataset, maxFeatures, maxSamples) method and its ResolvedParams import
  - IsolationForest.scala — removed private method, updated import and call site
  - extended/ExtendedIsolationForest.scala — removed private method, updated import and call site
